Make new friends! 🥳
fix - autoprefixer: It parses CSS and adds vendor prefixes to CSS rules using values from https://caniuse.com/. It is recommended by Google and used by Twitter and Taobao.
refactor - assets folder: Directory of assets such as images, css or javascript in a single folder
style - link as buttons: Highlighting of the real Yubo logo, without removing the Yubo title which is not visible on the page but existing for SEO,
change of the style of the links to buttons that make it cooler than before and that matches the style of the Yubo logo.
Obviously it's responsive and compatible with all browsers.